fix scratch/build.gradle: our ../api/src is (now) input
Strictly speaking just api/src/main/typescript.. but let's keep api/src;
just in case we in the future ever add e.g. src/js or something like
that.  At the worst, it will rebuild the scratch JS/TS also in case of
Java API changes - that's not the worst.  We currently don't need it,
but if we ever switch back to using the gen. sources, then we probably
would anyway, so this seems better because it's safer.