Do NOT mount host's ~/.gradle/ into build container after all
because this seems to cause conflicts between Gradle running
inside the build container and Gradle daemons running on the host
(e.g. in the Eclipse IDE), see e.g. the following (and similar) errors:
https://github.com/OASIS-learn-study/minecraft-storeys-maker/pull/381

This DOES mean that the Gradle Wrapper JAR is re-downloaded for
EVERY build, which I was trying to avoid, as it seems kind of stupid.
Surprisingly however, this doesn't seem to make the build much slower.